refactor(backend/drive): move Chunker inside
